use longs instead of ints, and minor wigismo runtime fix
git-svn-id: svn+ssh://svn.cs.mcgill.ca/2011/groups/group-d/wig@12689 6669c6d1-5a39-0410-bbc6-daf4d3d1134a

+++ src/tree.c
@@ -566,7 +566,7 @@ EXP *makeEXPcall(char *name, EXP *exps)
     return e;
 }
 
-EXP *makeEXPintconst(int i)
+EXP *makeEXPintconst(long i)
 {
     EXP *e = NEW(EXP);
     e->lineno = lineno;
